World Value Survey (Extract)

A cleaned version of the wvs dataset.

wvs_nona

Format

A data frame with 69578 rows and 6 variables:

country

Country name, partially abbreviated

gender

Gender of participant

age

Age of participant at the time of data collection

relationship_status

Relationship status of participants

freedom_of_choice

Item: Some people feel they have completely free choice and control over their lives, while other people feel that what they do has no real effect on what happens to them. Indicate how much freedom of choice and control you feel you have over the way your life turns out | 1 ("no choice at all") to 10 ("a great deal of choice").

satisfaction

Item: All things considered, how satisfied are you with your life as a whole these days? | 1 (“completely dissatisfied”) to 10 (“completely satisfied”)

Source

World Value Survey (https://www.worldvaluessurvey.org)