Global Education Profiler (random sample)

A dataset containing information about students' social integration (si) experience at their Higher Education Institutions and their communication skills (cs) development. For cs, scores of 'actualy experience' and 'importance to me' are available. This dataset only contains a random sample of 300 observations out of well over 12,000 students as of August 2021.

Usage

gep

Format

A data frame with 300 rows and 12 variables:

age

Age of participant

gender

Gender of participant

level_of_study

Degree level at the time of the research project

si_socialise_with_people_exp

Extent to which participants have opportunities to socialise with people from diverse cultural backgrounds | 1 ("not at all") to 6 ("very much")

si_supportive_friends_exp

Extent to which participants have supportive friends with a diverse cultural background | 1 ("not at all") to 6 ("very much")

si_time_socialising_exp

Extent to which participants take time to socialise with people from diverse cultural backgrounds | 1 ("not at all") to 6 ("very much")

cs_explain_ideas_imp

Extent to which participants perceive improving in explaining ideas to others (importance) | 1 ("not at all") to 6 ("very much")

cs_find_clarification_imp

Extent to which participants are able to clarify what others mean (importance) | 1 ("not at all") to 6 ("very much")

cs_learn_different_styles_imp

Extent to which participants learn about other students communication styles (importance) | 1 ("not at all") to 6 ("very much")

cs_explain_ideas_exp

Extent to which participants perceive improving in explaining ideas to others (experience) | 1 ("not at all") to 6 ("very much")

cs_find_clarification_exp

Extent to which participants are able to clarify what others mean (experience) | 1 ("not at all") to 6 ("very much")

cs_learn_different_styles_exp

Extent to which participants learn about other students communication styles (experience) | 1 ("not at all") to 6 ("very much")

Source

Global Education Profiler (https://warwick.ac.uk/gep)