Install R packages relevant for the book 'R for Non-Programmers: A Guide for Social Scientists'.

Install R packages relevant for the book 'R for Non-Programmers: A Guide for Social Scientists'.

install_r4np()

Details

Installs the following R packages at once:

broom: Turns output from functions into tibbles.

car: Interpret regression results

correlation: Lightweight package for computing different kinds of correlations, such as partial correlations, Bayesian correlations, multilevel correlations, polychoric correlations, biweight correlations, distance correlations and more.

effectsize: Provide utilities to work with indices of effect size and standardized parameters for a wide variety of models, allowing computation of and conversion between indices such as Cohen's d, r, odds, etc.

janitor: The main janitor functions can: perfectly f ormat data.frame column names; provide quick counts of variable combinations (i.e., frequency tables and crosstabs); and isolate duplicate records.

jtools: This is a collection of tools that the author (Jacob) has written for the purpose of more efficiently understanding and sharing the results of (primarily) regression analyses.

lavaan: Fit a variety of latent variable models, including confirmatory factor analysis, structural equation modeling and latent growth curve models.

mice: Multiple imputation using Fully Conditional Specification (FCS) implemented by the MICE algorithm as described in Van Buuren and Groothuis-Oudshoorn (2011).

mi: Provides functions for data manipulation, imputing missing values in an approximate Bayesian framework, diagnostics of the models used to generate the imputations, etc.

naniar: Provides data structures and functions that facilitate the plotting of missing values and examination of imputations.

parameters: Returns parameters from regression models.

patchwork: Combine different ggplots into one plot.

performance: Returns performance indices for regression models

psych: A general purpose toolbox for personality, psychometric theory and experimental psychology.

rcompanion: Additional functions, such as for effect sizes.

rstatix: A pipe-friendly way of computing various statistics.

skimr: A simple to use summary function that can be used with pipes and displays nicely in the console.

tidyverse: Install the core tidyverse packages.

Examples

if (FALSE) {
# Run this line of code to install all relevant packages
install_r4np()
}